Création d'un dossier utils pour séparer les fonctions utilitaires et éviter les imports circulaires
