+++ proxy_pda/utils/date_to_timezone.py
@@ -1,0 +1,7 @@
+from django.utils import timezone
+import datetime
+
+
+def date_to_timezone(date: str):
+    return timezone.make_aware(
+        datetime.datetime.strptime(date, '%Y-%m-%d %H:%M:%S'))
